// src/shared/enums/common.enum.ts
export enum Position {
  LEFT = 'left',
  RIGHT = 'right',
}

export enum Answer {
  TRUE = 'Tak',
  FALSE = 'Nie',
}

export enum FormatTyp {
  Bold = 'Bold',
  Boolean = 'Boolean',
  Currency = 'Currency',
  CurrencyAbs = 'CurrencyAbs',
  CurrencyGreater = 'CurrencyGreater',
  Currency6 = 'Currency6',
  Default = 'Default',
  Description = 'Description',
  GrayBoldTitle = 'GrayBoldTitle',
  Label = 'Label',
  LabelMedium = 'LabelMedium',
  LabelMargin = 'LabelMargin',
  LabelSmallMargin = 'LabelSmallMargin',
  LabelGreater = 'LabelGreater',
  Link = 'Link',
  MarginBottom4 = 'MarginBottom4',
  MarginBottom8 = 'MarginBottom8',
  MarginTop4 = 'MarginTop4',
  MarginTop8 = 'MarginTop8',
  Value = 'Value',
  ValueMedium = 'ValueMedium',
  GrayTitle = 'GrayTitle',
  HeaderContent = 'HeaderContent',
  SubHeaderContent = 'SubHeaderContent',
  TitleContent = 'TitleContent',
  HeaderPosition = 'HeaderPosition',
  Right = 'Right',
  DateTime = 'DateTime',
  Date = 'Date',
  FormOfPayment = 'FormOfPayment',
  Percentage = 'Percentage',
}

export default FormatTyp;